+++ src/compas_timber/connections/joint.py
@@ -221,13 +221,11 @@ class Joint(Interaction):
             return cls.promote_joint_candidate(model, cluster.joints[0], reordered_elements=elements, **kwargs)
         else:
             elements = reordered_elements or list(cluster.elements)
-            for joint in cluster.joints:
-                model.remove_joint(joint)
         return cls.create(model, *elements, **kwargs)
 
     @classmethod
     def promote_joint_candidate(cls, model, candidate, reordered_elements=None, **kwargs):
-        """Creates an instance of this joint from a generic joint.
+        """Creates an instance of this joint from a joint candidate.
 
         Parameters
         ----------
@@ -252,7 +250,7 @@ class Joint(Interaction):
             elements = reordered_elements
         else:
             elements = candidate.elements
-        model.remove_joint(candidate)
+        kwargs.update({"topology": candidate.topology, "location": candidate.location})  # pass topology, distance and location from candidate
         joint = cls.create(model, *elements, **kwargs)
         return joint